requirements.txt
flask==3.0.2
python-telegram-bot==20.8
SQLAlchemy==2.0.27
flask-sqlalchemy==3.1.0
flask-migrate==4.0.5
pyngrok==7.1.5
python-dotenv==1.0.1
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
Pillow==11.0.0
requests==2.31.0
gunicorn==21.2.0
PyJWT==2.8.0
gspread==6.2.0
oauth2client==4.1.3 